Tools to support proc_macros in Rust
FunctionDecl
-----------
Read into the FunctionDecl and then use the different struct parameters
to access different parts of the FuncitonDecl in order to print out the
desired re-written function body.

Uses a Regex to parse the function declaration and split it into
multiple sub-parts and then provide functions to re-piece them back
together with a new body.